# Conflicts:
#	csp_config.php

+++ src/pages/admin/actualites/supprimer.php
@@ -4,28 +4,36 @@ include(__DIR__ . '/../../core/connection.php');
 include(__DIR__ . '/../../../classes/actualite.php');
 
 
+require_once(__DIR__ . '/../../../../csp_config.php');
+
 
 $id = $_GET['id'];
-$query = "SELECT * FROM actualite WHERE id= $id";
-$result = $db->query($query);
-while ($row = $result->fetch_assoc()) {
+
+// Requête pour obtenir les informations de l'actualité
+$stmt = $db->prepare("SELECT * FROM actualite WHERE id = ?");
+$stmt->execute([$id]);
+$row = $stmt->fetch(PDO::FETCH_ASSOC);
+
+if ($row) {
     $image_path = __DIR__ . '/../../../../images/actualites/' . $row['image'];
-    $actualite = new actualite($db);
+
     // Vérifier si le fichier existe avant de le supprimer
     if (file_exists($image_path)) {
         unlink($image_path); // Supprimer le fichier
-        // Ajoutez ici toute autre logique nécessaire, comme mettre à jour la base de données, etc.
-        if ($actualite->delete($id)) {
-            header("Location: actualites.php");
-        }
-    } else {
+    }else {
         echo "L'image n'existe pas ou a déjà été supprimée.";
     }
-    // Affichage des images existantes avec une option de suppression
+        // Supprimer l'actualité de la base de données
+        $actualite = new actualite($db);
+        if ($actualite->delete($id)) {
+            header("Location: actualites.php");
+            exit(); // Arrêter l'exécution du script après la redirection
+        } else {
+            echo "Erreur lors de la suppression de l'actualité dans la base de données.";
+        }
 
-    $images_directory = __DIR__ . '/../../../../images/actualites/';
-    $images = scandir($images_directory);
-
+} else {
+    echo "Aucune actualité trouvée avec l'identifiant $id.";
 }
